Lots of changes and tweaks
Better formatting on some AI files

Added more aliases for Props

Updated the PlugAPI to sync with the most current iteration

Added manual UpdateCode.  Plug changed how it was done, so the old
static method of getting it automatically doesn't work until AtomJack
fixes the code.

Began support for a user tracker with numUsers and the onJoin and
onLeave events

Removed unnecessary logging

Added the ability to talk to the bot like a human with the
parseIncomingMessage() function and it's associated baggage.  This is
still a bit buggy and in need of refactoring, but it works for the most
part.  Relevant items are:  pareseIncomingMessage(), messageObject,
marvins_aliases, ALLOWED_COMMANDS