Final Check-In
Not sure how viable Marvin is anymore. Nobody has been to plug in days.
Final code push so that if someone else wants to pick this up (or if I
want to pick it back up) the last bit of progress won't be lost. Note
that this is untested and may not work.